Final integration of the Buffer button
This release should add the Buffer button to all existing setups.

For new installs, it should set up the floating bar with buttons as set out and styled at bufferapp.com/diggdigg

A lot of code tidying and reorganisation for convenience.

Removal of ads and references to previous dev/owner

Tidied up a lot of the language strings to read as better English

Cleaned up the readme (removed old version history, tidied up the tags etc)